requirements.txt
torch
torchvision
transformers
minigpt4
fastapi
uvicorn
pdf2image
pillow
python-docx
pytesseract